homepage functionality
Working through adding functionality to homepage. COmpleted the featured section, which required adding an 'updated_date' field to several models that auto sets itself to the current date when an item is saved. Changing the collection refferenced in the featured requires changing the view itself, need to change this moving forward so that non-base users (Person, etc. ) can manage